pyproject.toml
[tool.poetry]
name = "restic-backup"
version = "0.1.0"
description = ""
authors = ["Kyle Prestel <kprestel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"
yagmail = {extras = ["all"], version = "^0.11.224"}
systemd-python = "^234"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.782"
flake8 = "^3.8.3"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
restic-backup = 'restic_backup.backup:main'
